requirements.txt
fastapi
uvicorn
sqlmodel # ORM
psycopg2-binary # Just like python-mysql-connector
python-dotenv

# google-auth
# google-auth-oauthlib
# google-api-python-client
# youtube-transcript-api

httpx
google-api-python-client
google-auth
python-dotenv